+++ gd-sim/src/Vehicle.cpp
@@ -3,6 +3,7 @@
 #include <Object.hpp>
 #include <algorithm>
 #include <Slope.hpp>
+#include <cfloat>
 #include <cmath>
 
 /*
@@ -14,6 +15,7 @@ constexpr double velocity_thresholds[] = {
 	101.541492,
 	103.485494592,
 	103.377492,
+	103.809492,
 	103.809492
 };
 
@@ -45,17 +47,15 @@ Vehicle cube() {
 	Vehicle v;
 	 v.type = VehicleType::Cube;
 
-	v.enter = +[](Player& p, Object const*, bool n) {
-		if (n) {
-			if (p.prevPlayer().vehicle.type != VehicleType::Ball)
-				p.velocity = p.velocity / 2;
+	v.enter = +[](Player& p) {
+		if (p.prevPlayer().vehicle.type != VehicleType::Ball)
+			p.velocity = p.velocity / 2;
 
-			if (p.prevPlayer().vehicle.type == VehicleType::Ship && p.input)
-				p.buffer = true;
-		}
+		if (p.prevPlayer().vehicle.type == VehicleType::Wave)
+			p.velocity = p.velocity / 2;
 
-		p.ceiling = 999999;
-		p.floor = 0;
+		if (p.prevPlayer().vehicle.type == VehicleType::Ship && p.input)
+			p.buffer = true;
 	};
 
 	v.clamp = +[](Player& p) {
@@ -72,6 +72,7 @@ Vehicle cube() {
 			-2747.52,
 			-2794.1082, 
 			-2786.4,
+			-2799.36,
 			-2799.36
 		};
 		p.acceleration = accelerations[p.speed];
@@ -106,20 +107,12 @@ Vehicle cube() {
 				603.7217172,
 				616.681728,
 				606.421728,
+				606.421728
 			};
 
 			// On slopes, you jump higher depending on how long you've been on the slope
 			if (p.slopeData.slope && p.slopeData.slope->orientation == 0) {
 				auto time = std::clamp(10 * (p.timeElapsed - p.slopeData.elapsed), 0.4, 1.0);
-
-				/*static double slopeHeights[4] = {
-					322.345224,
-					399.889818,
-					497.224926,
-					600.643296
-				};*/
-
-
 				double vel = 0.9 * std::min(1.12 / p.slopeData.slope->angle(), 1.54) * (p.slopeData.slope->size.y * player_speeds[p.speed] / p.slopeData.slope->size.x);
 				p.setVelocity(0.25 * time * vel + jumpHeights[p.speed], p.prevPlayer().input);
 
@@ -132,6 +125,8 @@ Vehicle cube() {
 		}
 	};
 
+	v.bounds = FLT_MAX;
+
 	return v;
 }
 
@@ -139,16 +134,11 @@ Vehicle ship() {
 	Vehicle v;
 	v.type = VehicleType::Ship;
 
-	v.enter = +[](Player& p, Object const* o, bool n) {
-		if (n) {
-			if (p.prevPlayer().vehicle.type == VehicleType::Ufo || p.prevPlayer().vehicle.type == VehicleType::Wave)
-				p.velocity = p.velocity / 4.0;
-			else 
-				p.velocity = p.velocity / 2.0;
-		}
-
-		p.floor = std::max(0., std::ceil((o->pos.y - 180) / 30.)) * 30;
-		p.ceiling = p.floor + 300;
+	v.enter = +[](Player& p) {
+		if (p.prevPlayer().vehicle.type == VehicleType::Ufo || p.prevPlayer().vehicle.type == VehicleType::Wave)
+			p.velocity = p.velocity / 4.0;
+		else 
+			p.velocity = p.velocity / 2.0;
 	};
 
 	v.clamp = +[](Player& p) {
@@ -156,16 +146,18 @@ Vehicle ship() {
 		p.buffer = false;
 
 		// Max velocity
-		/*p.velocity = std::clamp(p.velocity, 
+		p.velocity = std::clamp(p.velocity, 
 			p.small ? -406.566 : -345.6,
 			p.small ? 508.248 : 432.0
-		);*/
+		);
+
+		//TODO deal with slope velocity more consistently
 
 		// Slopes complicate things like "maximum velocity"
-		if (p.input)
+		/*if (p.input)
 			p.velocity = std::min(p.velocity, p.small ? 508.248 : 432.0);
 		else
-			p.velocity = std::max(p.velocity, p.small ? -406.566 : -345.6);
+			p.velocity = std::max(p.velocity, p.small ? -406.566 : -345.6);*/
 
 
 		if (p.gravTop(p) > p.gravCeiling()) {
@@ -201,6 +193,8 @@ Vehicle ship() {
 		rotateFly(p, 0.15f);
 	};
 
+	v.bounds = 300;
+
 	return v;
 }
 
@@ -216,28 +210,27 @@ Vehicle ball() {
 
 		if (p.grav(p.pos.y) >= p.gravCeiling() && p.velocity > 0) {
 			p.setVelocity(0, true);
+
+			if (p.input)
+				p.upsideDown = !p.upsideDown;
 		}
 	};
 
-	v.enter = +[](Player& p, Object const* o, bool n) {
-		if (n) {
-			if (p.input)
-				p.vehicleBuffer = true;
+	v.enter = +[](Player& p) {
+		if (p.input)
+			p.vehicleBuffer = true;
 
-			switch (p.prevPlayer().vehicle.type) {
-				case VehicleType::Ship:
-				case VehicleType::Ufo:
-					p.velocity = p.velocity / 2;
-					break;
-				default: break;
-			}
+		switch (p.prevPlayer().vehicle.type) {
+			case VehicleType::Ship:
+			case VehicleType::Ufo:
+				p.velocity = p.velocity / 2;
+				break;
+			default: break;
 		}
-		p.floor = std::max(0., std::ceil((o->pos.y - 150) / 30.)) * 30;
-		p.ceiling = p.floor + 240;
 	};
 
 	v.update = +[](Player& p) {
-		if (!p.prevPlayer().velocityOverride)
+		if (!p.prevPlayer().velocityOverride || p.prevPlayer().slopeData.slope)
 			p.acceleration = -1676.46672;
 
 		if (!p.input)
@@ -262,6 +255,7 @@ Vehicle ball() {
 				-172.044007,
 				-181.11601,
 				-185.00401,
+				-181.92601,
 				-181.92601
 			};
 
@@ -283,6 +277,8 @@ Vehicle ball() {
 		}
 	};
 
+	v.bounds = 240;
+
 	return v;
 }
 
@@ -290,17 +286,12 @@ Vehicle ufo() {
 	Vehicle v;
 
 	v.type = VehicleType::Ufo;
-	v.enter = +[](Player& p, Object const* o, bool n) {
-		if (n) {
-			VehicleType pv = p.prevPlayer().vehicle.type;
-			if ((pv == VehicleType::Ship || pv == VehicleType::Wave) && p.input)
-				p.buffer = true;
+	v.enter = +[](Player& p) {
+		VehicleType pv = p.prevPlayer().vehicle.type;
+		if ((pv == VehicleType::Ship || pv == VehicleType::Wave) && p.input)
+			p.buffer = true;
 
-			p.velocity = p.velocity / (p.prevPlayer().vehicle.type == VehicleType::Ship ? 4 : 2);
-		}
-
-		p.floor = std::max(0., std::ceil((o->pos.y - 180) / 30.)) * 30;
-		p.ceiling = p.floor + 300;
+		p.velocity = p.velocity / (p.prevPlayer().vehicle.type == VehicleType::Ship ? 4 : 2);
 	};
 
 	v.clamp = +[](Player& p) {
@@ -342,30 +333,34 @@ Vehicle ufo() {
 		}
 	};
 
+	v.bounds = 300;
+
 	return v;
 }
 
 Vehicle wave() {
 	Vehicle v;
 	v.type = VehicleType::Wave;
-	v.enter = +[](Player& p, Object const* o, bool) {
+	v.enter = +[](Player& p) {
 		p.actions.push_back(+[](Player& p) {
 			p.size = p.small ? Vec2D(6, 6) : Vec2D(10, 10);
 		});
-
-		p.floor = std::max(0., std::ceil((o->pos.y - 180) / 30.)) * 30;
-		p.ceiling = p.floor + 300;
 	};
 
 	v.clamp = +[](Player& p) {
-		p.velocity = 0;
+		float waveTop = p.grav(p.pos.y + p.grav(p.size.y)) ;
+		float waveBottom = p.grav(p.pos.y - p.grav(p.size.y));
 
-		if (p.grav(p.pos.y - p.grav(p.size.y)) <= p.gravFloor() && !p.input) {
+		p.velocity = (p.input * 2 - 1) * player_speeds[p.speed] * (p.small ? 2 : 1);
+
+		if (waveBottom <= p.gravFloor()) {
 			p.pos.y = p.grav(p.gravFloor() + p.size.y);
-		} else if (p.grav(p.pos.y + p.grav(p.size.y)) >= p.gravCeiling() && p.input) {
+			if (waveBottom == p.gravFloor() && !p.input)
+				p.velocity = 0;
+		} else if (waveTop >= p.gravCeiling()) {
 			p.pos.y = p.grav(p.gravCeiling() - p.size.y);
-		} else {
-			p.velocity = (p.input * 2 - 1) * player_speeds[p.speed] * (p.small ? 2 : 1);
+			if (waveTop == p.gravCeiling() && p.input)
+				p.velocity = 0;
 		}
 
 		//p.rotation = (p.input ? 1 : -1) * (p.small ? 63.4258423 : 45);
@@ -376,6 +371,8 @@ Vehicle wave() {
 
 		rotateFly(p, p.small ? 0.4 : 0.25);
 	};
+
+	v.bounds = 300;
 
 	return v;
 }