+++ changelog.md
@@ -1,3 +1,15 @@
+# beta 22.1 (bugfix)
+- Fixed bug where player's hitbox was larger than intended
+
+# beta 22
+
+- Rotated object support for ship and wave
+- Slope fixes
+- 4x speed
+- Wave fixes
+- Debug button added
+- Fixed bounds for vehicles in level settings
+
 # beta 21.1 (minor fix)
 - Fixed bug where hitting pads while wave caused an error